fix: Move api hooks into folder
