+++ apps/dashboard/src/forms/ir_signup/ir_registration.page.tsx
@@ -1,7 +1,7 @@
 import { Button } from "@/components/ui/button"
 import { Checkbox } from "@/components/ui/checkbox"
-import { useDashboard } from "@/shared/hooks/useDashboard"
-import { useDates } from "@/shared/hooks/useDates"
+import { useDashboard } from "@/shared/hooks/api/useDashboard"
+import { useDates } from "@/shared/hooks/api/useDates"
 import { HOST } from "@/shared/vars"
 import { queryClient } from "@/utils/query_client"
 import { useMutation } from "@tanstack/react-query"
