+++ apps/dashboard/src/screens/dashboard/FormCard.tsx
@@ -1,8 +1,8 @@
 import { Badge } from "@/components/ui/badge"
 import { FORMS } from "@/forms"
 import { Form } from "@/forms/form_types"
-import { useDashboard } from "@/shared/hooks/useDashboard"
-import { useProducts } from "@/shared/hooks/useProducts"
+import { useDashboard } from "@/shared/hooks/api/useDashboard"
+import { useProducts } from "@/shared/hooks/api/useProducts"
 import { setActiveForm } from "@/store/form/form_slice"
 import { cx } from "@/utils/cx"
 import { useNavigate } from "@tanstack/react-router"
