+++ apps/dashboard/src/screens/dashboard/screen.tsx
@@ -10,8 +10,8 @@ import { ContactBubble } from "@/screens/dashboard/ContactBubble"
 import { getTimelinePhaseMessage } from "@/screens/dashboard/timeline_steps"
 import { LogoutButton } from "@/shared/LogoutButton"
 import { Timeline } from "@/shared/Timeline"
-import { useDashboard } from "@/shared/hooks/useDashboard"
-import { useDates } from "@/shared/hooks/useDates"
+import { useDashboard } from "@/shared/hooks/api/useDashboard"
+import { useDates } from "@/shared/hooks/api/useDates"
 import { selectForms } from "@/store/form/form_selectors"
 import { cx } from "@/utils/cx"
 import LoadingAnimation from "@/utils/loading_animation/loading_animation"
