+++ apps/dashboard/src/screens/form/screen.tsx
@@ -5,7 +5,7 @@ import { isFormOpen } from "@/forms/form_access"
 import { IfProgressDone } from "@/shared/IfProgressDone"
 import { InfoScreen } from "@/shared/InfoScreen"
 import { Navbar } from "@/shared/Navbar"
-import { useDashboard } from "@/shared/hooks/useDashboard"
+import { useDashboard } from "@/shared/hooks/api/useDashboard"
 import { remoteSaveChanges } from "@/store/form/async_actions"
 import { AppDispatch } from "@/store/store"
 import { useFormMeta } from "@/useFormMeta"
