+++ apps/dashboard/src/screens/form/sidebar/PageCard.tsx
@@ -1,8 +1,8 @@
 import { FormIds, FormPageIds } from "@/forms"
 import { Form, FormPage } from "@/forms/form_types"
 import { IfProgressDone } from "@/shared/IfProgressDone"
-import { useDashboard } from "@/shared/hooks/useDashboard"
-import { useProducts } from "@/shared/hooks/useProducts"
+import { useDashboard } from "@/shared/hooks/api/useDashboard"
+import { useProducts } from "@/shared/hooks/api/useProducts"
 import { remoteSaveChanges } from "@/store/form/async_actions"
 import { AppDispatch } from "@/store/store"
 import { cx } from "@/utils/cx"
