+++ apps/dashboard/src/shared/hooks/api/useRegistrationGroups.tsx
